add code to delete blog
only deletes blog if the decoded token gives a user id that matches the blog user id property

+++ controllers/blogs.js
@@ -69,8 +69,21 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    const result = await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(user.id.toString() === blog.user.toString()){
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+    else{
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
   }
   catch(exception) {
     next(exception)
